Fix: new fees adapter for paradex
